+++ spec/acceptance/16_dns_master_node_clean_deactivate.rb
@@ -76,7 +76,7 @@ EOS
         its(:exit_status) { is_expected.to eq 0 }
         its(:stdout) do
           is_expected.to match(
-            %r{sns.dns.icann.org. noc.dns.icann.org. 1 7200 3600 1209600 3600}
+            %r{sns.dns.icann.org. noc.dns.icann.org. 1 7200 3600 1209600 3600},
           )
         end
       end
@@ -84,7 +84,7 @@ EOS
         its(:exit_status) { is_expected.to eq 0 }
         its(:stdout) do
           is_expected.to match(
-            %r{sns.dns.icann.org. noc.dns.icann.org. 1 7200 3600 1209600 3600}
+            %r{sns.dns.icann.org. noc.dns.icann.org. 1 7200 3600 1209600 3600},
           )
         end
       end