+++ spec/acceptance/21_dns_virtualbox_master_notify_nsd_knot_spec.rb
@@ -80,7 +80,7 @@ EOS
         its(:exit_status) { is_expected.to eq 0 }
         its(:stdout) do
           is_expected.to match(
-            %r{sns.dns.icann.org. noc.dns.icann.org. 1 7200 3600 1209600 3600}
+            %r{sns.dns.icann.org. noc.dns.icann.org. 1 7200 3600 1209600 3600},
           )
         end
       end
@@ -88,7 +88,7 @@ EOS
         its(:exit_status) { is_expected.to eq 0 }
         its(:stdout) do
           is_expected.to match(
-            %r{sns.dns.icann.org. noc.dns.icann.org. 1 7200 3600 1209600 3600}
+            %r{sns.dns.icann.org. noc.dns.icann.org. 1 7200 3600 1209600 3600},
           )
         end
       end
@@ -104,7 +104,7 @@ EOS
         its(:exit_status) { is_expected.to eq 0 }
         its(:stdout) do
           is_expected.to match(
-            %r{sns.dns.icann.org. noc.dns.icann.org. 2 7200 3600 1209600 3600}
+            %r{sns.dns.icann.org. noc.dns.icann.org. 2 7200 3600 1209600 3600},
           )
         end
       end
@@ -112,7 +112,7 @@ EOS
         its(:exit_status) { is_expected.to eq 0 }
         its(:stdout) do
           is_expected.to match(
-            %r{sns.dns.icann.org. noc.dns.icann.org. 2 7200 3600 1209600 3600}
+            %r{sns.dns.icann.org. noc.dns.icann.org. 2 7200 3600 1209600 3600},
           )
         end
       end